requirements.txt
boto3>=1.4.7
botocore>=1.8.26
PyYAML==3.12
invoke==0.14.0
Jinja2==2.8
pynt==0.8.1
demjson==2.2.3
dnspython==1.15.0
jsonschema==2.6.0
deepdiff==3.3.0
